// frontend/src/api/generated/runtime.ts
export type FetchAPI = WindowOrWorkerGlobalScope['fetch'];
// ...
export interface RequestContext {
    fetch: FetchAPI;
    url: string;
    init: RequestInit;
}